refactor: update TypeOfOpHelper to include compile options in typeOfLogic calls
